server/requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.0.0
requests>=2.31.0
python-jose>=3.3.0
cryptography>=41.0.0 